+++ R/psychobject.R
@@ -1,0 +1,6 @@
+#' Creates or tests for objects of mode "psychobject".
+#'
+#' @param x an arbitrary R object.
+#'
+#' @export
+is.psychobject <- function(x) inherits(x, "psychobject")