+++ tests/testthat/test-analyze.glmerMod.R
@@ -1,9 +1,9 @@
 context("analyze.glmerMod")
 
 test_that("If it works.", {
-  # Fit
   require(lme4)
 
+  # GLM
   fit <- lme4::glmer(vs ~ mpg + (1 | cyl), data = mtcars, family = "binomial")
 
   model <- psycho::analyze(fit)